Create 1482. Minimum Number of Days to Make m Bouquets
Algorithm:
If m*k > arr.size: This means we have insufficient flowers. So, it is impossible to make m bouquets and we will return -1.
Next, we will find the maximum element i.e. max(arr[]), and the minimum element i.e. min(arr[]) in the array.
Place the 2 pointers i.e. low and high: Initially, we will place the pointers. The pointer low will point to min(arr[]) and the high will point to max(arr[]).
Calculate the ‘mid’: Now, inside the loop, we will calculate the value of ‘mid’ using the following formula:
mid = (low+high) // 2 ( ‘//’ refers to integer division)
Eliminate the halves based on the value returned by possible():
We will pass the potential answer, represented by the variable 'mid' (which corresponds to a specific day), to the 'possible()' function.
If possible() returns true: On satisfying this condition, we can conclude that the number ‘mid’ is one of our possible answers. But we want the minimum number. So, we will eliminate the right half and consider the left half(i.e. high = mid-1).
Otherwise, the value mid is smaller than the number we want. This means the numbers greater than ‘mid’ should be considered and the right half of ‘mid’ consists of such numbers. So, we will eliminate the left half and consider the right half(i.e. low = mid+1).
Finally, outside the loop, we will return the value of low as the pointer will be pointing to the answer.

Signed-off-by: Dhanasri Parisam <23MH1A4248@acoe.edu.in>