Create 1283. Find the Smallest Divisor Given a Threshold
Algorithm:
If n > threshold: If the minimum summation i.e. n > threshold value, the answer does not exist. In this case, we will return -1.
Next, we will find the maximum element i.e. max(arr[]) in the given array.
Place the 2 pointers i.e. low and high: Initially, we will place the pointers. The pointer low will point to 1 and the high will point to max(arr[]).
Calculate the ‘mid’: Now, inside the loop, we will calculate the value of ‘mid’ using the following formula:
mid = (low+high) // 2 ( ‘//’ refers to integer division)
Eliminate the halves based on the summation of division results:
We will pass the potential divisor, represented by the variable 'mid', to the 'sumByD()' function. This function will return the summation result of the division values.
If result <= threshold: On satisfying this condition, we can conclude that the number ‘mid’ is one of our possible answers. But we want the minimum number. So, we will eliminate the right half and consider the left half(i.e. high = mid-1).
Otherwise, the value mid is smaller than the number we want. This means the numbers greater than ‘mid’ should be considered and the right half of ‘mid’ consists of such numbers. So, we will eliminate the left half and consider the right half(i.e. low = mid+1).
Finally, outside the loop, we will return the value of low as the pointer will be pointing to the answer.

Signed-off-by: Dhanasri Parisam <23MH1A4248@acoe.edu.in>