Create **Allocate Books**
Algorithm:
If m > n: In this case, book allocation is not possible and so, we will return -1.
Next, we will find the maximum element and the summation of the given array.
We will use a loop(say pages) to check all possible pages from max(arr[]) to sum(arr[]).
Next, inside the loop, we will send each ‘pages’, to the function countStudents() function to get the number of students to whom we can allocate the books.
The first number of pages, ‘pages’, for which the number of students will be equal to ‘m’, will be our answer. So, we will return that particular ‘pages’.
Finally, if we are out of the loop, we will return max(arr[]) as there cannot exist any answer smaller than that.

Signed-off-by: Dhanasri Parisam <23MH1A4248@acoe.edu.in>