Update 383. Ransom Note
The first character is 'h'. We check mp['h'], which is 1, so we can use it. We decrement mp['h'] to 0. Now, the map is: mp = {'h': 0, 'e': 1, 'o': 2, 'l': 2}.
The second character is 'e'. We check mp['e'], which is 1, so we can use it. We decrement mp['e'] to 0. Now, the map is: mp = {'h': 0, 'e': 0, 'o': 2, 'l': 2}.
The third character is 'l'. We check mp['l'], which is 2, so we can use it. We decrement mp['l'] to 1. Now, the map is: mp = {'h': 0, 'e': 0, 'o': 2, 'l': 1}.
The fourth character is 'l' again. We check mp['l'], which is 1, so we can use it. We decrement mp['l'] to 0. Now, the map is: mp = {'h': 0, 'e': 0, 'o': 2, 'l': 0}.
The fifth character is 'o'. We check mp['o'], which is 2, so we can use it. We decrement mp['o'] to 1. Now, the map is: mp = {'h': 0, 'e': 0, 'o': 1, 'l': 0}.
Final Result:

All characters of ransomNote = "hello" have been matched and used from magazine. After the loop, the function returns true because we could construct the ransom note from the available characters in the magazine.

Signed-off-by: Dhanasri Parisam <23MH1A4248@acoe.edu.in>