Rename **Count the number of subarrays with given xor K** to Count the number of subarrays with given xor K
Approach
Prefix XOR: We can calculate the XOR of elements up to each position in the array and use this to determine the XOR of any subarray efficiently.

Using a HashMap: We'll use a hash map to store the frequency of each prefix XOR encountered so far. This helps in quickly finding subarrays with the desired XOR.

Logic:

For each element in the array, calculate the cumulative XOR from the start of the array up to the current element (let's call this prefixXOR).
For a subarray ending at the current index to have an XOR equal to k, there must exist a prefix XOR (prefixXOR ^ k) at some earlier position.
Check if prefixXOR ^ k exists in the hash map. If it does, the value at that key in the map will be the number of subarrays ending at the current index with XOR equal to k.
Update the hash map by adding the current prefixXOR.

Signed-off-by: Dhanasri Parisam <23MH1A4248@acoe.edu.in>